bulk-cdk-toolkit-extract-jdbc: move Constants to jdbc module
