+++ airbyte-cdk/bulk/toolkits/extract-jdbc/src/main/kotlin/io/airbyte/cdk/read/JdbcPartitionsCreatorFactory.kt
@@ -5,6 +5,7 @@
 package io.airbyte.cdk.read
 
 import io.airbyte.cdk.command.OpaqueStateValue
+import io.airbyte.cdk.jdbc.JDBC_PROPERTY_PREFIX
 import io.micronaut.context.annotation.Requires
 import jakarta.inject.Singleton
 
@@ -68,6 +69,4 @@ class JdbcConcurrentPartitionsCreatorFactory<
         JdbcConcurrentPartitionsCreator(partition, partitionFactory)
 }
 
-const val JDBC_PROPERTY_PREFIX = "airbyte.connector.extract.jdbc"
-
-private const val MODE_PROPERTY = "$JDBC_PROPERTY_PREFIX.mode"
+private const val MODE_PROPERTY = "${JDBC_PROPERTY_PREFIX}.mode"
